Fixed the get-cef script for windows.

+++ get-cef.ps1
@@ -5,30 +5,32 @@ $CEF_ARCHIVE = "cef_binary_122.1.12+g6e69d20+chromium-122.0.6261.112_windows64_m
 $ErrorActionPreference = "Stop"
 
 
-mkdir -f hoi
+mkdir -f cef
 if (!(Test-Path "cef\$CEF_ARCHIVE.tar.bz2")) {
 	"Downloading CEF..."
-	#$client = new-object System.Net.WebClient
-	#$client.DownloadFile("https://cef-builds.spotifycdn.com/$CEF_ARCHIVE.tar.bz2", "cef\$CEF_ARCHIVE.tar.bz2")
 	curl -o "cef\$CEF_ARCHIVE.tar.bz2.part" "https://cef-builds.spotifycdn.com/$CEF_ARCHIVE.tar.bz2"
 	mv "cef\$CEF_ARCHIVE.tar.bz2.part" "cef\$CEF_ARCHIVE.tar.bz2"
 }
 
-"Unpacking CEF..."
-tar -xvf "cef\$CEF_ARCHIVE.tar.bz2" -C cef
+if (!(Test-Path "cef\$CEF_ARCHIVE")) {
+	"Unpacking CEF..."
+	tar -xvf "cef\$CEF_ARCHIVE.tar.bz2" -C cef
+}
 
 "Compiling CEF..."
 try {
 	cd "cef\$CEF_ARCHIVE"
 
 	# Add compilation definitions to the top of the CMakeLists.txt file
-	mv CMakeLists.txt CMakeLists.txt.old
-	[IO.File]::WriteAllLines("CMakeLists.txt", "add_compile_definitions(DCHECK_ALWAYS_ON=1)")
-	$FROM = Get-Content -Path "CMakeLists.txt.old"
-	Add-Content -Path "CMakeLists.txt" -Value $FROM
+	if (!(Test-Path "CMakeLists.txt.def")) {
+		mv CMakeLists.txt CMakeLists.txt.old
+		Set-Content -Path "CMakeLists.txt.def" -Value "add_compile_definitions(NDEBUG=1 DCHECK_ALWAYS_ON=1)"
+		Get-Content CMakeLists.txt.def, CMakeLists.txt.old | Set-Content -Path "CMakeLists.txt"
+	}
 
 	cmake .
-	cmake --build .
+	cmake --build . --config Release
+	cmake --build libcef_dll_wrapper --config Release --target libcef_dll_wrapper
 }
 finally {
 	cd ..\..
